+++ js/data-error.js
@@ -1,0 +1,1 @@
+// Модуль, который реагирует на неправильно введенные данные
