+++ js/data-success.js
@@ -1,0 +1,1 @@
+// Модуль, который показывает сообщение об успешной отправке данных
