+++ js/send-data.js
@@ -1,0 +1,1 @@
+// Модуль, который отправляет информацию на сервер при нажатии на кнопку "Опубликовать"
